+++ src/components/Weight/Weight.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import '../Weight/Weight.css';
+
+const Weight = () => {
+    return (
+        <div className='placeholder-view fade-in'>
+            <div className='placeholder-content'>
+                <h1>Weight Tracking</h1>
+                    <p>Feature coming soon.</p>
+            </div>
+        </div>
+    );
+};
+
+export default Weight;